feat: user cannot delete himself

+++ backend/src/user/service/user.service.ts
@@ -1,10 +1,11 @@
-import { Injectable } from '@nestjs/common';
+import { ForbiddenException, Injectable } from '@nestjs/common';
 import { Not } from 'typeorm';
 
 import { RegisterUserInput } from 'src/auth/dto/input/register.user.input';
 
 import { UserRepository } from '../repository/user.repository';
 import { PaginationInput } from './../dto/input/pagination.input';
+import Lang from 'src/constants/language';
 
 @Injectable()
 export class UserService {
@@ -17,7 +18,11 @@ export class UserService {
     );
   }
 
-  async deleteUserById(userId: string) {
+  async deleteUserById(userId: string, deleterId: string) {
+    if (userId === deleterId) {
+      throw new ForbiddenException(Lang.CANNOT_DELETE_YOURSELF);
+    }
+
     return await this.userRepository.deleteById(userId);
   }
